@@ -1,4 +1,4 @@
 export const LOCATION_CHANGE = '@@router/LOCATION_CHANGE'
-// I'm not sure, but I guess that if you use redux-devtools extension your APP_INIT should be look like:
+// XXX: DONT USE INTERNAL REDUX STUFF! e.g. @@INIT is an internal action
 // export const APP_INIT = '@@INIT'
-export const APP_INIT = '@@redux/INIT'
+export const APPLICATION_INIT = 'APPLICATION_INIT'
