@@ -1,5 +1,5 @@
 export * from './auth'
 export * from './common'
-export * from './dashboard'
-export * from './inbox'
+export * from './posts'
+export * from './users'
 export * from './layout'
