@@ -1,4 +1,5 @@
 export * from './AuthSvc'
-export * from './InboxSvc'
-export * from './StatisticsSvc'
+export * from './LocalStorageCookiesSvc'
+export * from './PostsSvc'
+export * from './UsersSvc'
 export * from './utils'
