@@ -1,5 +1,2 @@
-export SidebarComponent from './Sidebar'
-export Header from './Header'
-export Footer from './Footer'
-export Root from './Root'
-export RouteAuth from './RouteAuth'
+export * from './views'
+export * from './addons'
