@@ -1,178 +1,244 @@
 import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, matchPath} from 'react-router'
-// import {Helmet} from 'react-helmet'
-// Accessing PropTypes via the main React package is deprecated.
-// Use the prop-types package from npm instead.
 import PropTypes from 'prop-types'
 import {push} from 'react-router-redux'
-import {Dimmer, Sidebar, Container} from 'semantic-ui-react'
-import {Header, SidebarComponent, Footer} from 'components'
+// Import main views
+import Sidebar from 'components/views/Sidebar'
+import Footer from 'components/views/Footer'
+import Header from 'components/views/Header'
+// Import actions
 import {CLOSE_SIDEBAR, OPEN_SIDEBAR, WINDOW_RESIZE} from 'actions/layout'
 import {LOGOUT_AUTH} from 'actions/auth'
-import {appRouting} from 'routing'
-import './App.scss'
+// Import styled components
+import {
+	PageLayout,
+	MainLayout,
+	MainContent,
+	SidebarSemanticPusherStyled,
+	SidebarSemanticPushableStyled,
+	MainContainer,
+	StyledDimmer
+} from './style'
 
 class App extends Component {
-  static propTypes = {
-    children: PropTypes.node.isRequired,
-    // react-router `withRouter` props
-    location: PropTypes.object,
-    history: PropTypes.object,
+	static propTypes = {
+		children: PropTypes.node.isRequired,
+		// Routes of app passed as props in `Root`
+		routes: PropTypes.array.isRequired,
+		// React-router `withRouter` props
+		location: PropTypes.object,
+		history: PropTypes.object,
 
-    // sidebarOpened can force component to re-render
-    sidebarOpened: PropTypes.bool,
-    closeSidebar: PropTypes.func,
-    // isLoggedIn can force component to re-render
-    isLoggedIn: PropTypes.bool,
-    handleWindowResize: PropTypes.func,
-    logout: PropTypes.func,
-    checkAuthLogic: PropTypes.func,
-    toggleSidebar: PropTypes.func,
-    // isMobile can force component to re-render
-    isMobile: PropTypes.bool
-  }
+		// SidebarOpened can force component to re-render
+		sidebarOpened: PropTypes.bool,
+		closeSidebar: PropTypes.func,
+		// IsLoggedIn can force component to re-render
+		isLoggedIn: PropTypes.bool,
+		handleWindowResize: PropTypes.func,
+		logout: PropTypes.func,
+		checkAuthLogic: PropTypes.func,
+		toggleSidebar: PropTypes.func,
+		// IsMobile can force component to re-render
+		isMobile: PropTypes.bool,
+		isMobileXS: PropTypes.bool,
+		isMobileSM: PropTypes.bool
+	}
 
-  // XXX: must be fixed.
-  // shouldComponentUpdate(nextProps) {
-  //     let {match, isMobile, isLoggedIn, sidebarOpened} = this.props
-  //     let matchSame = _.isEqual(nextProps.match, match)
-  //     let isMobileSame = _.isEqual(nextProps.isMobile, isMobile)
-  //     let isLoggedInSame = _.isEqual(nextProps.isLoggedIn, isLoggedIn)
-  //     let sidebarOpenedSame = _.isEqual(nextProps.sidebarOpened, sidebarOpened)
-  //     // return props that can force us aren't the same
-  //     return !(matchSame && isMobileSame && isLoggedInSame && sidebarOpenedSame)
-  // }
+	// XXX: will be fixed one day.
+	// shouldComponentUpdate(nextProps) {
+	//     let {match, isMobile, isLoggedIn, sidebarOpened} = this.props
+	//     let matchSame = _.isEqual(nextProps.match, match)
+	//     let isMobileSame = _.isEqual(nextProps.isMobile, isMobile)
+	//     let isLoggedInSame = _.isEqual(nextProps.isLoggedIn, isLoggedIn)
+	//     let sidebarOpenedSame = _.isEqual(nextProps.sidebarOpened, sidebarOpened)
+	//     // return props that can force us aren't the same
+	//     return !(matchSame && isMobileSame && isLoggedInSame && sidebarOpenedSame)
+	// }
 
-  componentWillMount () {
-    let {handleWindowResize, isLoggedIn} = this.props
-    if (process.env.BROWSER) {
-      window.addEventListener('resize', handleWindowResize)
-    }
-    this.checkAppAuthLogic(isLoggedIn)
-  }
+	componentWillMount () {
+		const {isLoggedIn} = this.props
+		if (process.env.BROWSER) {
+			const {handleWindowResize} = this.props
+			window.addEventListener('resize', handleWindowResize)
+		}
+		this.checkAppAuthLogic(isLoggedIn)
+	}
 
-  /**
+	/**
+   * Checks that user is still allowed to visit path after props changed
+   * @param  {Object} nextProps
+   */
+	componentWillReceiveProps (nextProps) {
+		this.checkAppAuthLogic(nextProps.isLoggedIn)
+	}
+
+	componentDidMount () {
+		if (process.env.SENTRY_PUBLIC_DSN) {
+			const script = document.createElement('script')
+			script.type = 'text/javascript'
+			script.crossorigin = 'anonymous'
+			script.async = true
+			script.onload = () => {
+				Raven.config(process.env.SENTRY_PUBLIC_DSN).install()
+			}
+			script.src = 'https://cdn.ravenjs.com/3.16.1/raven.min.js'
+			document.body.appendChild(script)
+		}
+
+		if (process.env.GA_ID) {
+			const script = document.createElement('script')
+			script.type = 'text/javascript'
+			script.async = true
+			script.crossorigin = 'anonymous'
+			script.onload = () => {
+				window.ga =
+					window.ga ||
+					function () {
+						(ga.q = ga.q || []).push(arguments)
+					}
+				ga.l = Number(new Date())
+				ga('create', process.env.GA_ID, 'auto')
+				ga('send', 'pageview')
+			}
+			script.src = 'https://www.google-analytics.com/analytics.js'
+			document.body.appendChild(script)
+		}
+	}
+
+	/**
      * Check that user is allowed to visit route
-     * @param  {Bool} loggedIn state.auth.loggedIn, current prop
-     * @return {Null} Nothing
+     * @param  {Bool} isLoggedIn state.auth.me.isLoggedIn, current prop
+     * @return {Undefined} Nothing
      */
-  checkAppAuthLogic (loggedIn) {
-    const {location, checkAuthLogic} = this.props
-    const path = location.pathname
-    checkAuthLogic(path, loggedIn)
-  }
+	checkAppAuthLogic (isLoggedIn) {
+		const {location, checkAuthLogic} = this.props
+		const path = location.pathname
+		checkAuthLogic(path, isLoggedIn)
+	}
 
-  componentWillReceiveProps (nextProps) {
-    this.checkAppAuthLogic(nextProps.isLoggedIn)
-  }
+	/**
+   * Returns routing for sidebar menu
+   * @return {Array} array of routes that will be rendered in sidebar menu
+   */
+	getSidebarRouting () {
+		const sidebarRouting = this.props.routes.filter(a => a.sidebarVisible).map(a => {
+			const {path, name, icon, external, strict, exact} = a
+			const b = {path, name, icon, external, strict, exact}
+			return b
+		})
+		return sidebarRouting
+	}
 
-  getSidebarRouting () {
-    // routing for sidebar menu
-    const sidebarRouting = appRouting.filter(a => a.sidebarVisible).map(a => {
-      const {path, name, icon, external, strict, exact} = a
-      const b = {path, name, icon, external, strict, exact}
-      return b
-    })
-    return sidebarRouting
-  }
-
-  /**
-  * returns title for header
+	/**
+  * Returns title for header
   * @param  {String} pathname - location.pathname
   * @return {String} page title
   */
-  getPageTitle (pathname) {
-    const matchedRoutes = appRouting.filter(a => matchPath(pathname, a))
-    const currentRoute = matchedRoutes[0] || {}
-    const title = currentRoute.name || '404'
-    return title
-  }
+	getPageTitle (pathname) {
+		const matchedRoutes = this.props.routes.filter(a => matchPath(pathname, a))
+		const currentRoute = matchedRoutes[0] || {}
+		const title = currentRoute.name || '404'
+		return title
+	}
 
-  render () {
-    const {
-      children,
-      sidebarOpened,
-      closeSidebar,
-      isLoggedIn,
-      logout,
-      toggleSidebar,
-      location,
-      isMobile
-    } = this.props
+	render () {
+		const {
+			children,
+			sidebarOpened,
+			closeSidebar,
+			isLoggedIn,
+			logout,
+			toggleSidebar,
+			location,
+			isMobile
+		} = this.props
 
-    // routing for sidebar menu
-    const sidebarRouting = this.getSidebarRouting()
-    const title = this.getPageTitle(location.pathname)
+		// Routing for sidebar menu
+		const sidebarRouting = this.getSidebarRouting()
+		const title = this.getPageTitle(location.pathname)
 
-    const sidebarProps = {
-      isMobile,
-      logout,
-      open: sidebarOpened,
-      routing: sidebarRouting
-    }
+		const sidebarProps = {
+			isMobile,
+			logout,
+			open: sidebarOpened,
+			routing: sidebarRouting
+		}
 
-    const headerProps = {
-      toggleSidebar,
-      title,
-      isLoggedIn
-    }
+		const headerProps = {
+			toggleSidebar,
+			title,
+			isLoggedIn,
+			isMobile
+		}
 
-    const dimmerProps = {
-      active: true,
-      onClick: closeSidebar
-    }
+		const dimmerProps = {
+			//  Dimmed: true,
+			active: isLoggedIn && sidebarOpened,
+			page: true,
+			//  Blurring: true,
+			//  page: true,
+			onClick: closeSidebar
+		}
 
-    return (
-      <div className="page-layout">
-        <Sidebar.Pushable>
-          {isLoggedIn && <SidebarComponent {...sidebarProps} />}
-          <Sidebar.Pusher>
-            {/* Semantic ui currently(16.04.16) doesn't have closeDimmerOnClick or smth else
-            So, instead of it, we can use simple <Dimmer> component */}
-            {/* <SidebarSemantic.Pusher dimmed={sidebarOpened}> */}
-            <Header {...headerProps} />
-            <main>
-              <div className="main-content">
-                <Container>
-                  {children}
-                </Container>
-              </div>
-              <Footer />
-            </main>
-            {/* show dimmer only if:
-              1. isLoggedIn, elsewhere sidebar isn't visible
-            2. if sidebar is opened  */}
-            {isLoggedIn && sidebarOpened && <Dimmer {...dimmerProps} />}
-          </Sidebar.Pusher>
-        </Sidebar.Pushable>
+		// {/* XXX: There is an issue with props and styled-components, so we use .extend and re-render the component when isMobile/isLoggedIn change triggered. Using `style` attribute isn't a good solution.
+		//   Please, check: https://github.com/styled-components/styled-components/issues/439 */}
+		//   {/* <SidebarSemanticPusherStyled style={SidebarSemanticPusherStyleProps}> */}
+		const SidebarSemanticPusherStyledPatch =
+			!isMobile && isLoggedIn
+				? SidebarSemanticPusherStyled.extend`
+						max-width: calc(100% - 150px);
+					`
+				: SidebarSemanticPusherStyled
 
-      </div>
-    )
-  }
+		return (
+			<PageLayout>
+				<SidebarSemanticPushableStyled>
+					{isLoggedIn && <Sidebar {...sidebarProps}/>}
+					<SidebarSemanticPusherStyledPatch>
+						<StyledDimmer {...dimmerProps}/>
+						<Header {...headerProps}/>
+						<MainLayout>
+							<MainContent>
+								<MainContainer id="main-container">
+									{children}
+								</MainContainer>
+							</MainContent>
+							<Footer/>
+						</MainLayout>
+						{/* </Dimmer.Dimmable> */}
+					</SidebarSemanticPusherStyledPatch>
+				</SidebarSemanticPushableStyled>
+			</PageLayout>
+		)
+	}
 }
 
 function mapStateToProps (state) {
-  return {
-    sidebarOpened: state.layout.sidebarOpened,
-    isMobile: state.layout.isMobile,
-    isLoggedIn: state.auth.loggedIn
-  }
+	const {sidebarOpened, isMobile, isMobileXS, isMobileSM} = state.layout
+	const {isLoggedIn} = state.me.auth
+	return {
+		sidebarOpened,
+		isMobile,
+		isMobileXS,
+		isMobileSM,
+		isLoggedIn
+	}
 }
 
 function mapDispatchToProps (dispatch) {
-  let resizer
-  return {
-    closeSidebar: () => {
-      dispatch(CLOSE_SIDEBAR())
-    },
-    logout: () => {
-      dispatch(LOGOUT_AUTH())
-    },
-    toggleSidebar: () => {
-      dispatch(OPEN_SIDEBAR())
-    },
-    /**
+	let resizer
+	return {
+		closeSidebar: () => {
+			dispatch(CLOSE_SIDEBAR())
+		},
+		logout: () => {
+			dispatch(LOGOUT_AUTH())
+		},
+		toggleSidebar: () => {
+			dispatch(OPEN_SIDEBAR())
+		},
+		/**
          * Immediately push to homePath('/'), if user is logged.
          * Can be used for other auth logic checks.
          * Useful, because we don't need to dispatch `push(homePath)` action
@@ -180,18 +246,18 @@ function mapDispatchToProps (dispatch) {
          * @param  {String}  path       [current location path]
          * @param  {Boolean} isLoggedIn [is user logged in?]
          */
-    checkAuthLogic: (path, isLoggedIn) => {
-      const authPath = '/auth'
-      const homePath = '/'
-      if (isLoggedIn && path === authPath) {
-        dispatch(push(homePath))
-      }
-    },
-    handleWindowResize: () => {
-      clearTimeout(resizer)
-      resizer = setTimeout(() => dispatch(WINDOW_RESIZE()), 100)
-    }
-  }
+		checkAuthLogic: (path, isLoggedIn) => {
+			const authPath = '/auth'
+			const homePath = '/'
+			if (isLoggedIn && path === authPath) {
+				dispatch(push(homePath))
+			}
+		},
+		handleWindowResize: () => {
+			clearTimeout(resizer)
+			resizer = setTimeout(() => dispatch(WINDOW_RESIZE()), 150)
+		}
+	}
 }
 
 export default withRouter(connect(mapStateToProps, mapDispatchToProps)(App))
