@@ -1,39 +1,40 @@
 import React, {Component} from 'react'
-import {Card, Loader, Grid} from 'semantic-ui-react'
+import {Card, Grid} from 'semantic-ui-react'
 import PropTypes from 'prop-types'
 import _ from 'lodash'
 import DashboardCardComponent from './DashboardCardComponent'
 
 export default class DashboardComponent extends Component {
-  static propTypes = {
-    statistics: PropTypes.array
-  }
+	static propTypes = {
+		posts: PropTypes.object,
+		postsLoaded: PropTypes.bool,
+		postsLoading: PropTypes.bool,
+		count: PropTypes.number
+	}
 
-  shouldComponentUpdate (nextProps) {
-    const {statistics} = this.props
-    const nextStatistics = nextProps.statistics
-    return !_.isEqual(statistics, nextStatistics)
-  }
+	shouldComponentUpdate (nextProps) {
+		const {posts} = this.props
+		const nextPosts = nextProps.posts
+		return !_.isEqual(posts, nextPosts)
+	}
 
-  render () {
-    const {statistics} = this.props
+	render () {
+		// {count, postsLoading}
+		const {posts, postsLoaded} = this.props
 
-    return (
-      <div>
-        {statistics.length === 0 && <Loader active={true}>Loading...</Loader>}
-        <Grid columns={1}>
-          <Grid.Row centered>
-            <Grid.Column width={16}>
-              {statistics &&
-                <Card.Group itemsPerRow={2} doubling>
-                  {statistics.map((stats, i) => (
-                    <DashboardCardComponent {...stats} key={i} />
-                  ))}
-                </Card.Group>}
-            </Grid.Column>
-          </Grid.Row>
-        </Grid>
-      </div>
-    )
-  }
+		return (
+			<Grid columns={1}>
+				<Grid.Row centered>
+					<Grid.Column width={16}>
+						{postsLoaded &&
+							<Card.Group itemsPerRow={3} doubling stackable>
+								{_.map(posts, (post, i) =>
+									<DashboardCardComponent {...post} key={i} />
+								)}
+							</Card.Group>}
+					</Grid.Column>
+				</Grid.Row>
+			</Grid>
+		)
+	}
 }
