@@ -1,4 +1,6 @@
 export App from './App'
 export Dashboard from './Dashboard'
-export Inbox from './Inbox'
+export Users from './Users'
 export Login from './Login'
+// NOTE: Export only containers that aren't lazy-loaded!
+// Else lazy containers would be imported in the main bundle!
